Replace device.DeviceProfile with device.ProfileName
Remove the embedded DeviceProfile struct from the Device.
The Device struct will keep just the ProfileName as a reference.
The Device SDK have a cache of Device Profiles, so they will be retrieved from the cache.
All other services will make a call to metadata in order to get a Device Profile.

It decreases the Device SDK cache size, and descrease the debug and error log messages.

Resolves/Fixes edgexfoundry/go-mod-core-contracts#27

Signed-off-by: difince <dianaa@vmware.com>